+++ IronNet/src/ironattack/decoy_scan.h
@@ -1,0 +1,6 @@
+#ifndef IRON_DECOY_SCAN_H
+#define IRON_DECOY_SCAN_H
+
+int cmd_decoy_scan(int argc, char **argv);
+
+#endif